fix: improve index pages template (#107)
Improve the decoration for index pages on categories and breeds.

Test URLs:
- Before: https://main--petplace--hlxsites.hlx.page/article/breed/
- After: https://index-pages--petplace--hlxsites.hlx.page/article/breed/
