+++ scripts/scripts.js
@@ -20,11 +20,11 @@ const LCP_BLOCKS = ['slideshow']; // add your LCP blocks to the list
 let templateModule;
 window.hlx.RUM_GENERATION = 'project-1'; // add your RUM generation information here
 
-function getId() {
+export function getId() {
   return Math.random().toString(32).substring(2);
 }
 
-function isMobile() {
+export function isMobile() {
   return window.innerWidth < 1024;
 }
 
@@ -74,54 +74,6 @@ export async function getCategoryByName(categoryName) {
   return categories.data.find((c) => c.Category === categoryName);
 }
 
-function buildCategorySidebar() {
-  const section = document.createElement('div');
-  section.classList.add('sidebar');
-  section.setAttribute('role', 'complementary');
-
-  const id1 = getId();
-  const id2 = getId();
-  const filterToggle = document.createElement('button');
-  filterToggle.disabled = !isMobile();
-  filterToggle.setAttribute('aria-controls', `${id1} ${id2}`);
-  filterToggle.textContent = 'Filters';
-  section.append(filterToggle);
-
-  const subCategories = buildBlock('sub-categories', { elems: [] });
-  subCategories.id = id1;
-  subCategories.setAttribute('aria-hidden', isMobile());
-  section.append(subCategories);
-
-  const popularTags = buildBlock('popular-tags', { elems: [] });
-  popularTags.id = id2;
-  popularTags.setAttribute('aria-hidden', isMobile());
-  section.append(popularTags);
-
-  filterToggle.addEventListener('click', () => {
-    const isVisible = subCategories.getAttribute('aria-hidden') === 'false';
-    if (!isVisible) {
-      filterToggle.dataset.mobileVisible = true;
-    }
-    subCategories.setAttribute('aria-hidden', isVisible);
-    popularTags.setAttribute('aria-hidden', isVisible);
-  });
-
-  window.addEventListener('resize', () => {
-    const isVisible = subCategories.getAttribute('aria-hidden') === 'false';
-    if (!isVisible && !isMobile()) {
-      filterToggle.disabled = true;
-      subCategories.setAttribute('aria-hidden', false);
-      popularTags.setAttribute('aria-hidden', false);
-    } else if (isVisible && isMobile() && !filterToggle.dataset.mobileVisible) {
-      filterToggle.disabled = false;
-      subCategories.setAttribute('aria-hidden', true);
-      popularTags.setAttribute('aria-hidden', true);
-    }
-  }, { passive: true });
-
-  return section;
-}
-
 /**
  * Builds hero block and prepends to main in a new section.
  * @param {Element} main The container element
@@ -137,6 +89,7 @@ async function buildHeroBlock(main) {
   }
   const h1 = main.querySelector('h1');
   const picture = main.querySelector('picture');
+  const parent = h1.parentElement;
   // eslint-disable-next-line no-bitwise
   if (h1 && picture && (h1.compareDocumentPosition(picture) & Node.DOCUMENT_POSITION_PRECEDING)) {
     const img = picture.querySelector('img');
@@ -151,6 +104,7 @@ async function buildHeroBlock(main) {
       section.append(buildBlock('hero', { elems: [optimized, h1] }));
     }
     main.prepend(section);
+    parent.remove();
   }
 }
 
@@ -208,14 +162,8 @@ async function decorateTemplate(main) {
  * @param {Element} main The container element
  */
 async function buildAutoBlocks(main) {
-  const bodyClass = [...document.body.classList];
-
   try {
     await buildHeroBlock(main);
-
-    if (bodyClass.includes('category-index')) {
-      main.insertBefore(buildCategorySidebar(), main.querySelector(':scope > div:nth-of-type(2)'));
-    }
   } catch (error) {
     // eslint-disable-next-line no-console
     console.error('Auto Blocking failed', error);
