+++ templates/breed-index/breed-index.css
@@ -39,6 +39,50 @@
     max-width: 520px;
 }
 
+.breed-index .sidebar {
+  padding: 1rem;
+  background-color: var(--background-color-dark);
+}
+
+.breed-index .sidebar > .default-content-wrapper:first-child {
+  text-align: center;
+}
+
+.breed-index .sidebar button[aria-controls] {
+  margin: 0;
+  background: none;
+  color: var(--highlight-color);
+  font-size: var(--body-font-size-l);
+}
+
+.breed-index .sidebar button[aria-controls]::before {
+  display: inline-block;
+  margin-right: 0.5rem;
+  content: "\2715";
+  font-size: 22px;
+  transform: rotate(45deg);
+  transition: transform 1s linear;
+}
+
+.breed-index .sidebar h2 {
+  position: relative;
+  padding-bottom: 6px;
+  border-bottom: 1px solid var(--color-gray);
+  color: var(--text-color-inverted);
+  font-size: var(--heading-font-size-l);
+}
+
+.breed-index .sidebar h2::after {
+  position: absolute;
+  content: '';
+  display: block;
+  bottom: -2px;
+  width: 15%;
+  min-width: 100px;
+  max-width: 200px;
+  border-bottom: 3px solid var(--color-purple);
+}
+
 @media screen and (min-width: 461px) {
     .hero-wrapper .img-div {
         position: relative;
@@ -50,6 +94,47 @@
     .hero-wrapper .text-div {
         top: 60px
     }
+}
+
+@media (min-width: 1024px) {
+  .breed-index main {
+    display: grid;
+    grid-template:
+      'hero hero hero hero' min-content
+      '. sidebar content .' 1fr
+      '. sidebar pagination .' min-content
+      'articles articles articles articles' min-content / 1fr minmax(250px, 342px) minmax(750px, 1026px) 1fr;
+    gap: 0 1.5rem;
+  }
+
+  .breed-index .hero-container {
+    grid-area: hero;
+  }
+
+  .breed-index .sidebar {
+    grid-area: sidebar;
+    background: unset;
+  }
+
+  .breed-index .pagination-container {
+    grid-area: pagination;
+  }
+
+  .breed-index .cards-container{
+    grid-area: content;
+  }
+
+  .breed-index .slide-cards-container {
+    grid-area: articles;
+  }
+
+  .breed-index .sidebar h2 {
+    color: unset;
+  }
+
+  .breed-index .sidebar button[aria-controls] {
+    display: none;
+  }
 }
 
 @media (min-width: 1200px) {
