+++ templates/category-index/category-index.js
@@ -1,6 +1,67 @@
-import { getCategory } from '../../scripts/scripts.js';
+import { buildBlock } from '../../scripts/lib-franklin.js';
+import { getCategory, getId, isMobile } from '../../scripts/scripts.js';
 
-// eslint-disable-next-line import/prefer-default-export
+function buildSidebar() {
+  const section = document.createElement('div');
+  section.classList.add('sidebar');
+  section.setAttribute('role', 'complementary');
+
+  const id1 = getId();
+  const id2 = getId();
+  const filterToggle = document.createElement('button');
+  filterToggle.disabled = !isMobile();
+  filterToggle.setAttribute('aria-controls', `${id1} ${id2}`);
+  filterToggle.textContent = 'Filters';
+  section.append(filterToggle);
+
+  const subCategories = buildBlock('sub-categories', { elems: [] });
+  subCategories.id = id1;
+  subCategories.setAttribute('aria-hidden', isMobile());
+  section.append(subCategories);
+
+  const popularTags = buildBlock('popular-tags', { elems: [] });
+  popularTags.id = id2;
+  popularTags.setAttribute('aria-hidden', isMobile());
+  section.append(popularTags);
+
+  filterToggle.addEventListener('click', () => {
+    const isVisible = subCategories.getAttribute('aria-hidden') === 'false';
+    if (!isVisible) {
+      filterToggle.dataset.mobileVisible = true;
+    }
+    subCategories.setAttribute('aria-hidden', isVisible);
+    popularTags.setAttribute('aria-hidden', isVisible);
+  });
+
+  window.addEventListener('resize', () => {
+    const isVisible = subCategories.getAttribute('aria-hidden') === 'false';
+    if (!isVisible && !isMobile()) {
+      filterToggle.disabled = true;
+      subCategories.setAttribute('aria-hidden', false);
+      popularTags.setAttribute('aria-hidden', false);
+    } else if (isVisible && isMobile() && !filterToggle.dataset.mobileVisible) {
+      filterToggle.disabled = false;
+      subCategories.setAttribute('aria-hidden', true);
+      popularTags.setAttribute('aria-hidden', true);
+    }
+  }, { passive: true });
+
+  return section;
+}
+
+function createTemplateBlock(main, blockName) {
+  const section = document.createElement('div');
+
+  const block = buildBlock(blockName, { elems: [] });
+  section.append(block);
+  main.append(section);
+}
+
+export function loadEager(main) {
+  main.insertBefore(buildSidebar(), main.querySelector(':scope > div:nth-of-type(2)'));
+  createTemplateBlock(main, 'pagination');
+}
+
 export async function loadLazy() {
   const heroColorDiv = document.querySelector('.category-index .hero > div');
   const path = window.location.pathname;
